+++ src/worldloom/connector_definition.py
@@ -37,6 +37,7 @@ REFERENCE_CONNECTORS = (
     "slack",
     "teamwork_graph",
     "rovo",
+    "sor",
 )
 
 ConnectorMaturity = Literal["ga", "beta", "eap", "product_surface"]